British Parliament Personal Wealth Data

Data set used in experiment to analyze the causal effect of being elected to Parliament in the United Kingdom on one's personal wealth.

MPs

Format

A data frame with 427 rows and 10 variables:

surname

character: surname of the candidate

firstname

character: first name of the candidate

party

character: party of the candidate (labour or tory)

ln.gross

numeric: log gross wealth at the time of death

ln.net

numeric: log net wealth at the time of death

yob

integer: year of birth of the candidate

yod

integer: year of death of the candidate

margin.pre

numeric: margin of the candidate’s party in the previous election

region

character: electoral region

margin

numeric: margin of victory (vote share)

Details

See QSS Table 4.8.

References

Imai, Kosuke. 2017. Quantitative Social Science: An Introduction. Princeton University Press. http://press.princeton.edu/titles/11025.html.

Andrew C. Eggers and Jens Hainmueller (2009) “MPs for sale? Returns to office in postwar British politics.” American Political Science Review, vol. 103, no. 4, pp. 513–533. doi = 10.1017/S0003055409990190