STAR Project Data

Contains data from the STAR (Student–Teacher Achievement Ratio) Project, a four-year longitudinal study examining the effect of class size in early grade levels on educational performance and personal development

STAR

Format

A data frame with 6325 rows and 6 variables:

race

integer: student’s race (white = 1, black = 2, Asian = 3, Hispanic = 4, Native American = 5, others = 6)

classtype

integer: type of kindergarten class (small = 1, regular = 2, regular with aid = 3)

yearssmall

integer: number of years in small classes

hsgrad

integer: high-school graduation (did graduate = 1, did not graduate = 0)

g4math

integer: total scaled score for the math portion of the fourth-grade standardized test

g4reading

integer: total scaled score for the reading portion of the fourth-grade standardized test

Details

See QSS Table 2.6.

References

Imai, Kosuke. 2017. Quantitative Social Science: An Introduction. Princeton University Press. URL.

Mosteller, Frederick. 1997. “The Tennessee Study of Class Size in the Early School Grades.” Bulletin of the American Academy of Arts and Sciences 50(7): 14-25. doi = 10.2307/3824562