Sex Ratio and the Price of Agricultural Crops in China

Demographic and agricultural crop data for individual counties in China.

chinawomen

Format

A data frame with 51766 rows and 9 variables:

admin

integer: unique county identifier

biryr

integer: year of cohort (birth year)

birpop

integer: birth population in a given year

han

numeric

sex

numeric: proportion of males in the birth cohort

teasown

numeric: quantity of tea sown in the county

orch

numeric: quantity of orchard-type crops planted in the county

cashcrop

numeric: quantity of cash crops planted in the county

post

integer: indicator variable for the introduction of price reforms

Details

See QSS Table 7.4.

References

Imai, Kosuke. 2017. Quantitative Social Science: An Introduction. Princeton University Press. http://press.princeton.edu/titles/11025.html.

Missing women and the price of tea in China: The effect of sex-specific earnings on sex imbalance.” Quarterly Journal of Economics, vol. 123, no. 3, pp. 1251–1285.