Facial Appearanace Experiment Data

Contains data from an experiment using competence measures (self-reported based on brief exposure to image of a candidate's face) to predict election outcomes.

face

Format

A data frame with 119 rows and 10 variables:

year

integer: year of the election

state

character: state of the election

winner

character: name of the winner

loser

character: name of the runner-up

w.party

character: party of the winner

l.party

character: party of the loser

d.comp

numeric: competence measure for the Democratic candidate

r.comp

numeric: competence measure for the Republican candidate

d.votes

integer: number of votes for the Democratic candidate

r.votes

integer: number of votes for the Republican candidate

Details

See QSS Table 4.4.

References

Imai, Kosuke. 2017. Quantitative Social Science: An Introduction. Princeton University Press. http://press.princeton.edu/titles/11025.html.

Alexander Todorov, Anesu N. Mandisodza, Amir Goren, and Crystal C. Hall (2005) “Inferences of competence from faces predict election outcomes.” Science, vol. 308, no. 10 (June), pp. 1623–1626. doi: 10.1126/science.1110589