Progresa Program Data

Data set from experiment which seeks to estimate the electoral impact of Progresa, Mexico’s conditional cash transfer program (CCT program).

progresa

Format

A data frame with 417 rows and 20 variables:

treatment

integer: whether an electoral precinct contains a village where households received early Progresa

pri2000s

numeric: PRI votes in the 2000 election as a share of precinct population above 18

pri2000v

numeric: official PRI vote share in the 2000 election

t2000

numeric: turnout in the 2000 election as a share of precinct population above 18

t2000r

numeric: official turnout in the 2000 election

pri1994

integer: total PRI votes in the 1994 presidential election

pan1994

integer: total PAN votes in the 1994 presidential election

prd1994

integer: total PRD votes in the 1994 presidential election

pri1994s

numeric: total PRI votes in the 1994 election as a share of precinct population above 18

pri1994v

numeric: official PRI vote share in the 1994 election

pan1994s

numeric: total PAN votes in the 1994 election as a share of precinct population above 18

pan1994v

numeric: official PAN vote share in the 1994 election

prd1994s

numeric: total PRD votes in the 1994 election as a share of precinct population above 18

prd1994v

numeric: official PRD vote share in the 1994 election

t1994

numeric: turnout in the 1994 election as a share of precinct population above 18

t1994r

numeric: official turnout in the 1994 election

votos1994

integer: total votes cast in the 1994 presidential election

avgpoverty

numeric: precinct average of village poverty index

pobtot1994

integer: total population in the precinct

villages

integer: number of villages in the precinct

Details

See QSS Table 4.11.

References

Imai, Kosuke. 2017. Quantitative Social Science: An Introduction. Princeton University Press. URL.

Ana de la O (2013) “Do conditional cash transfers affect voting behavior? Evidence from a randomized experiment in Mexico.” American Journal of Political Science, vol. 57, no. 1, pp. 1–14 doi: 10.1111/j.1540-5907.2012.00617.x

Kosuke Imai, Gary King, and Carlos Velasco (2015) “Do nonpartisan programmatic policies have partisan electoral effects? Evidence from two large scale randomized experiments.”