UN Voting Data

Dataset used to analyze how states’ ideological positions, as captured by their votes on UN resolutions, have changed since the fall of communism.

unvoting

Format

A data frame with 9120 rows and 6 variables:

Year

integer: year for which the ideal point is estimated

CountryAbb

character: abbreviated name of the country

CountryName

character: name of the country

idealpoint

numeric: its estimated ideal point

PctAgreeUS

numeric: proportion of votes that match with votes cast by the United States on the same issue

PctAgreeRUSSIA

numeric: proportion of votes that match with votes cast by Russia/the Soviet Union on the same issue

Details

See QSS Table 3.8.

References

Imai, Kosuke. 2017. Quantitative Social Science: An Introduction. Princeton University Press. http://press.princeton.edu/titles/11025.html.

Michael A. Bailey, Anton Strezhnev, and Erik Voeten (2015) “Estimating dynamic state preferences from United Nations voting data.” Journal of Conflict Resolution, doi = 10.1177/0022002715595700.