Women as Policy Makers Experiment

Data set from experiment analyzing the causal effects of having female politicians in government on policy outcomes

women

Format

A data frame with 322 rows and 6 variables:

GP

integer: identifier for the Gram Panchayat (GP)

village

integer: identifier for each village

reserved

integer: binary variable indicating whether the GP was reserved for women leaders or not

female

integer: binary variable indicating whether the GP had a female leader or not

irrigation

integer: variable measuring the number of new or repaired irrigation facilities in the village since the reserve policy started

water

integer: variable measuring the number of new or repaired drinking water facilities in the village since the reservation policy started

Details

See QSS Table 4.7.

References

Imai, Kosuke. 2017. Quantitative Social Science: An Introduction. Princeton University Press. URL.

Raghabendra Chattopadhyay and Esther Duflo (2004) “Women as policy makers: Evidence from a randomized policy experiment in India.” Econometrica, vol. 72, no. 5, pp. 1409–1443. doi: 10.1111/j.1468-0262.2004.00539.x